% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bes_2015}
\alias{bes_2015}
\title{A tibble with the British Election Study 2015 Constituency Results Version 2.2.}
\format{A tibble, containing 46 rows and 3 columns.
\describe{
}}
\source{
\url{http://www.britishelectionstudy.com/data-object/2015-bes-constituency-results-with-census-and-candidate-data/}
}
\usage{
bes_2015
}
\description{
A tibble with the British Election Study 2015 Constituency Results Version 2.2.
}
\references{
British Election Study 2015 Constituency Results Version 2.2. DOI: 10.13140/RG.2.1.1162.1844
}
\keyword{datasets}
