% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bes_2019}
\alias{bes_2019}
\title{2019 General Election Results}
\format{A tibble, containing 650 rows and 98 columns. For full details on
the included variables see the vignette:
\url{http://docs.evanodell.com/parlitools/articles/bes-2019.html}}
\source{
\url{https://candidates.democracyclub.org.uk/api/docs/csv/#past}

\url{https://researchbriefings.parliament.uk/ResearchBriefing/Summary/CBP-8749}
}
\usage{
bes_2019
}
\description{
A tibble with the 2019 UK General Election results, taken from the House
of Commons library and the Democracy Club crowd-sourced data on candidates.
}
\details{
A tibble with Great Britain constituencies results from the
2019 General Election. This data can be linked to 2011 census
information in \code{\link{census_11}} with the \code{pano},
\code{ons_const_id} and \code{constituency_name} variables.

Although not taken from the official British Election Study I have preserved
the \verb{bes_\{election\}} naming convention.
Data from Northern Ireland is also included in this object.
}
\seealso{
bes_2015

bes_2017

census_2011
}
\keyword{datasets}
