% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{council_data}
\alias{council_data}
\title{Constituency demographic information}
\format{A tibble, containing 632 rows and 191 columns. For full details see the vignette: \url{http://docs.evanodell.com/parlitools/articles/census-11.html}}
\source{
\url{http://www.britishelectionstudy.com/data-object/2017-bes-constituency-results-with-census-and-candidate-data/}
A tibble with details on the council make-up of local authorities across the UK.

This data is taken from \url{http://opencouncildata.co.uk/}, run by Jon Lawson. Variable names have been converted to snake_case and variables have been converted to appropriate R classes. ONS local authority codes have been taken from the \code{local_hex_map} data, and from the ONS's Open Geography Portal. This data is accurate as of 2017-06-02.

\url{http://opencouncildata.co.uk/councils.php?model=}
\url{http://geoportal.statistics.gov.uk/datasets/464be6191a434a91a5fa2f52c7433333_0}
}
\usage{
council_data
}
\description{
A tibble with demographic information on each constituency, taken from the 2011 census data included in the British Election Study. Can be linked to the \code{\link{bes_2017}} and \code{\link{bes_2015}} datasets using the \code{pano}, \code{ons_const_id} and \code{constituency_name} variables.
}
\keyword{datasets}
