% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{local_hex_map}
\alias{local_hex_map}
\title{Hexagonal cartogram, of all Local Authorities in England, Wales and Scotland. Originally downloaded from ESRI, with coordinates in Easting and Northing. Converted coordinates into Latitude and Longitude in QGIS.}
\format{An sf and data.frame object, containing 650 elements and 8 variables.
\describe{
  \item{object_id}{Row Name}
  \item{la_code}{Local Authority identifier code}
  \item{la_name}{Local Authority Name}
  \item{shape_length}{The lengths of the shape}
  \item{shape_area}{The area of the shape}
  \item{geometry}{Shape geometry/projection}
}}
\source{
\url{http://www.arcgis.com/home/item.html?id=593037bc399e460bb7c6c631ceff67b4}
}
\usage{
local_hex_map
}
\description{
Hexagonal cartogram, of all Local Authorities in England, Wales and Scotland. Originally downloaded from ESRI, with coordinates in Easting and Northing. Converted coordinates into Latitude and Longitude in QGIS.
}
\keyword{datasets}
