% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{west_hex_map}
\alias{west_hex_map}
\alias{local_hex_map}
\title{A hexagonal cartogram of Westminster parliamentary constituencies,
stored as a simple feature.}
\format{An sf and data.frame object, containing 650 elements and 8 variables.
\describe{
\item{\code{object_id}}{Row Name}
\item{\code{constituency_name}}{The name of the constituency}
\item{\code{description}}{Description of constituency type}
\item{\code{gss_code}}{The constituency GSS (Government Statistical
Service) code}
\item{\code{region_name}}{Name of region within the UK}
\item{\code{shape_length}}{The lengths of the shape}
\item{\code{shape_area}}{The area of shape}
\item{\code{geometry}}{Shape geometry/projection}
}}
\source{
\url{http://www.arcgis.com/home/item.html?id=15baaa6fecd54aa4b7250780b6534682}

\url{http://www.arcgis.com/home/item.html?id=593037bc399e460bb7c6c631ceff67b4}
}
\usage{
west_hex_map

local_hex_map
}
\description{
Originally created by Ben Flanagan at ESRI, with coordinates in Easting
and Northing. The coordinates in the original shapefile were converted
into Latitude and Longitude in QGIS, and the hexagon representing Ynys
Mon was moved to be closer to the mainland of Wales.

Originally created by Ben Flanagan at ESRI, with coordinates in Easting and
Northing. Converted coordinates into Latitude and Longitude in QGIS.
}
\keyword{datasets}
