% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_on_date.R
\name{mps_on_date}
\alias{mps_on_date}
\title{mps_on_date}
\usage{
mps_on_date(date1 = Sys.Date(), date2 = NULL, tidy = TRUE)
}
\arguments{
\item{date1}{The date to return the list of mps from. Defaults to current system date.}

\item{date2}{An optional query parameter. If a proper date in "YYYY-MM-DD" format, the function returns a list of all MPs who were members between date2 and date1. Defaults to NULL.}

\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.}
}
\value{
A tibble
}
\description{
mps_on_date
}
\examples{
\dontrun{

x <- mp_on_date()

}
}
