% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_ge_2017}
\alias{ni_ge_2017}
\title{Northern Ireland GE Results}
\format{A tibble, with 18 rows and 62 columns.}
\source{
\url{https://www.electoralcommission.org.uk/our-work/our-research/electoral-data/electoral-data-files-and-reports}
}
\usage{
ni_ge_2017
}
\description{
A tibble with results from UK General Elections in Northern Ireland
in 2015 and 2017.
}
\details{
Data is taken from the Electoral Commission results.
}
\keyword{datasets}
