% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlitools_tidy.R
\name{parlitools_tidy}
\alias{parlitools_tidy}
\title{parlitools_tidy}
\usage{
parlitools_tidy(df, tidy_style)
}
\arguments{
\item{df}{The tibble to tidy}

\item{tidy_style}{The style to tidy variable names to.}
}
\description{
A function to make the results of calls to the API easier to work with, mostly though fixing variable names, which by default contain non alpha-numeric characters and appear to use an inconsistent/idiosyncratic naming convention, at least by the standards of the various naming conventions used in R. Dates and datetimes are converted to POSIXct class. Some extra URL data included in the API is also stripped out.
}
