% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spending_2001}
\alias{spending_2001}
\title{Tibbles with the election spending of all candidates in the 2001 General Election. Variable names have been converted to snake_case.}
\format{A tibble with 3291 rows and 22 columns.
\describe{
}}
\source{
\url{http://www.electoralcommission.org.uk/find-information-by-subject/elections-and-referendums/past-elections-and-referendums/uk-general-elections/candidate-election-spending}
}
\usage{
spending_2001
}
\description{
Tibbles with the election spending of all candidates in the 2001 General Election. Variable names have been converted to snake_case.
}
\keyword{datasets}
