% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spending_2010}
\alias{spending_2010}
\title{A tibble with the election spending of all candidates in the 2010 General Election. Variable names have been converted to snake_case.}
\format{A tibble with 4031 rows and 52 columns.
\describe{
}}
\source{
\url{http://www.electoralcommission.org.uk/find-information-by-subject/elections-and-referendums/past-elections-and-referendums/uk-general-elections/candidate-election-spending}
}
\usage{
spending_2010
}
\description{
A tibble with the election spending of all candidates in the 2010 General Election. Variable names have been converted to snake_case.
}
\keyword{datasets}
