% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spending_2015}
\alias{spending_2015}
\title{Tibbles with the election spending of all candidates in the 2015 General Election. Variable names have been converted to snake_case.}
\format{A tibble with 3971 rows and 50 columns.
\describe{
}}
\source{
\url{http://www.electoralcommission.org.uk/find-information-by-subject/elections-and-referendums/past-elections-and-referendums/uk-general-elections/candidate-election-spending}
}
\usage{
spending_2015
}
\description{
Tibbles with the election spending of all candidates in the 2015 General Election. Variable names have been converted to snake_case.
}
\keyword{datasets}
