% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing.R
\name{swing}
\alias{swing}
\title{swing}
\usage{
swing(swing1 = c("Conservative", "Green", "Labour", "Liberal Democrat",
  "Plaid Cymru", "Scottish National Party", "UKIP"),
  swing2 = c("Conservative", "Green", "Labour", "Liberal Democrat",
  "Plaid Cymru", "Scottish National Party", "UKIP"), constituency = NULL,
  region = NULL, swing_style = c("Butler", "Steed"))
}
\arguments{
\item{swing1}{}

\item{swing2}{}

\item{constituency}{}

\item{region}{}

\item{swing_style}{}
}
\description{
swing
}
