% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{west_hex_map}
\alias{west_hex_map}
\title{A hexagonal cartogram of Westminster parliamentary constituencies}
\format{A SpatialPolygonsDataFrame, containing 650 elements
\describe{
}}
\source{
\url{}
}
\usage{
west_hex_map
}
\description{
A hexagonal cartogram of Westminster parliamentary constituencies
}
\keyword{datasets}
