% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westminster_seat_map.R
\name{westminster_seat_map}
\alias{westminster_seat_map}
\title{westminster_seat_map}
\usage{
westminster_seat_map(date = Sys.Date(), hex_map = FALSE)
}
\arguments{
\item{date}{The date to return data from}

\item{hex_map}{The map to use}
}
\value{
Returns a map of all seats in the UK Parliament
}
\description{
westminster_seat_map
}
\examples{


}
