% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bes_2015}
\alias{bes_2015}
\title{2015 General Election Results}
\format{
A tibble, containing 650 rows and 92 columns. For full details on
the included variables see the vignette:
\url{http://docs.evanodell.com/parlitools/articles/bes-2015.html}
}
\source{
\url{http://www.britishelectionstudy.com/data-object/2015-bes-constituency-results-with-census-and-candidate-data/}
}
\usage{
bes_2015
}
\description{
A tibble with the British Election Study
2015 Constituency Results Version 2.21.
}
\details{
A tibble with constituencies results from the British Election Study 2015.
Census data from the BES has been moved to the \code{\link{census_11}}
dataset, and can be linked through the with the \code{pano},
\code{ons_const_id} and \code{constituency_name} variables. Only contains
detailed results from Great Britain; the only data included on Northern
Irish constituencies is the party that won the seat.

Variable names have been converted to snake_case and variables have been
converted to appropriate R classes. If using in publications, please
cite as below:
}
\references{
Fieldhouse, Edward, Jane Green, Geoffrey Evans,
Hermann Schmitt, Cees van der Eijk, Jonathan Mellon, and
Christopher Prosser. "British Election Study, 2015: General Election
Results Dataset," 2015. doi:10.13140/RG.2.1.1162.1844.
}
\seealso{
bes_2017

bes_2019

census_2011
}
\keyword{datasets}
