% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bes_2015}
\alias{bes_2015}
\title{A tibble with the British Election Study 2015 Constituency Results Version 2.2. Variable names have been converted to snake_case and variables have been converted to appropriate R classes.}
\format{A tibble, containing 632 rows and 277 columns.}
\source{
\url{http://www.britishelectionstudy.com/data-object/2015-bes-constituency-results-with-census-and-candidate-data/}
}
\usage{
bes_2015
}
\description{
A tibble with the British Election Study 2015 Constituency Results Version 2.2. Variable names have been converted to snake_case and variables have been converted to appropriate R classes.
}
\references{
Fieldhouse, Edward, Jane Green, Geoffrey Evans, Hermann Schmitt, Cees van der Eijk, Jonathan Mellon, and Christopher Prosser. “British Election Study, 2015: General Election Results Dataset,” 2015. doi:10.13140/RG.2.1.1162.1844.
}
\keyword{datasets}
