% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/council_seats.R
\name{council_seats}
\alias{council_seats}
\title{Council Seats}
\usage{
council_seats(councillors = FALSE)
}
\arguments{
\item{councillors}{If \code{TRUE}, downloads details on each individual
councillor. Defaults to \code{FALSE} and downloads summary data for each council.}
}
\value{
A tibble with council seat distribution
}
\description{
Downloads data from
<\url{http://opencouncildata.co.uk/}>, with the
composition of each local council in Great Britain.
}
