% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_mps.R
\name{current_mps}
\alias{current_mps}
\title{All current MPs}
\usage{
current_mps(tidy = TRUE, tidy_style = "snake_case")
}
\arguments{
\item{tidy}{If \code{TRUE}, fixes the variable names in the tibble to remove 
special characters and superfluous text, and converts the variable names to 
a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if 
\code{tidy=TRUE}. Accepts one of \code{'snake_case'}, \code{'camelCase'} 
and \code{'period.case'}. Defaults to \code{'snake_case'}.}
}
\value{
A tibble of all MPs currently eligible to sit in the House of Commons.
}
\description{
Request data on all MPs eligible currently (i.e. on the current system date) 
to sit in the House of Commons. Includes information on their constituency.
}
\examples{
\dontrun{
x <- current_mps()
}
}
