% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_mps.R
\name{current_mps}
\alias{current_mps}
\title{current_mps}
\usage{
current_mps(tidy = TRUE)
}
\arguments{
\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.}
}
\value{
A tibble of all MPs currently eligible to sit in the House of Commons
}
\description{
current_mps
}
\examples{
\dontrun{

x <- current_mps()

}
}
