% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ge_2017}
\alias{ge_2017}
\title{A tibble with the results of the 2017 general election.}
\format{A tibble, containing 632 row and 245 columns. For details on all variables in the dataset please see the vignette: \url{http://docs.evanodell.com/parlitools/articles/ge-2017.html}}
\source{
\url{researchbriefings.parliament.uk/ResearchBriefing/Summary/CBP-7979} \url{http://www.britishelectionstudy.com/data-object/2015-bes-constituency-results-with-census-and-candidate-data/}
}
\usage{
ge_2017
}
\description{
A tibble with constituencies results from the 2017 general election, linked to census data for each constituency from the British Election Study 2015. Only covers Great Britain, so data on Northern Irish constituencies is not included.
}
\details{
Variable names have been converted to snake_case and variables have been converted to appropriate R classes.
}
\keyword{datasets}
