% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leave_votes_west}
\alias{leave_votes_west}
\title{A tibble with details on percentage of votes cast for leave in the 2016 referendum on British membershi of the European Union.}
\format{A tibble, with 641 rows and 7 columns.
\describe{
\item{gss_code}{ONS code for constituency}
\item{constituency}{The name of the constituency.}
\item{party_2016}{The party holding the constituency at the time of the referendum.}
\item{ch_leave_vote}{Estimates of the leave vote produced by Chris Hanretty.}
\item{known_leave_vote}{A dummy variable indicating 'yes' if the actual leave vote percentage is known, and 'no' if the estimate by CHris Hanretty is available.}
\item{known_leave_vote_perc}{The percentage of votes cast for leave, if known.}
\item{difference_estimate_known}{Difference between the estimate by Chris Hanretty and the actual result, if known.}
}}
\source{
\url{https://secondreading.uk/brexit/brexit-votes-by-constituency/}
}
\usage{
leave_votes_west
}
\description{
A tibble with details on percentage of votes cast for leave in the 2016 referendum on British membershi of the European Union.
}
\keyword{datasets}
