% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlitools-package.R
\docType{package}
\name{parlitools}
\alias{parlitools}
\title{parlitools: Tools for Analysing UK Politics with R}
\description{
The parlitools package provides various useful tools and functions for
analysing UK politics, creating political maps and retrieving data.
}
