% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{party_colour}
\alias{party_colour}
\title{A tibble with the ID, name and hex code for the official colour of a variety of political parties.}
\format{A tibble, containing 47 rows and 3 columns.
\describe{
\item{\code{party_id}}{The party ID, as assigned by the Members Name Information Service}
\item{\code{party_name}}{The name of the political party}
\item{\code{party_colour}}{A hex code for the party colour, as used by Wikipedia}
}}
\source{
\url{https://en.wikipedia.org/wiki/Wikipedia:Index_of_United_Kingdom_political_parties_meta_attributes}
\url{http://data.parliament.uk/membersdataplatform/services/mnis/ReferenceData/Parties/}
}
\usage{
party_colour
}
\description{
Includes all political parties with MPs and a number without MPs. It also contains a 'No Overall Control' party for local government. Hex codes are taken from Wikipedia.
}
\keyword{datasets}
