% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{west_hex_map}
\alias{west_hex_map}
\title{A hexagonal cartogram of Westminster parliamentary constituencies, stored as a simple feature. Originally downloaded from ESRI, with coordinates in Easting and Northing. The coordinates in the original shapefile were converted into Latitude and Longitude in QGIS, and the hexagon representing Ynys Mon was moved to be closer to the mainland of Wales.}
\format{An sf and data.frame object, containing 650 elements and 8 variables.
\describe{
  \item{object_id}{Row Name}
  \item{constituency_name}{The name of the constituency}
  \item{description}{Description of constituency type}
  \item{gss_code}{The constituency GSS (Government Statistical Service) code}
  \item{region_name}{Name of region within the UK}
  \item{shape_length}{The lengths of the shape}
  \item{shape_area}{The area of shape}
  \item{geometry}{Shape geometry/projection}
}}
\source{
\url{http://www.arcgis.com/home/item.html?id=15baaa6fecd54aa4b7250780b6534682}
}
\usage{
west_hex_map
}
\description{
A hexagonal cartogram of Westminster parliamentary constituencies, stored as a simple feature. Originally downloaded from ESRI, with coordinates in Easting and Northing. The coordinates in the original shapefile were converted into Latitude and Longitude in QGIS, and the hexagon representing Ynys Mon was moved to be closer to the mainland of Wales.
}
\keyword{datasets}
